code polish
