+++ starter/starter-configcenter/starter-zookeeper/src/main/java/org/dromara/dynamictp/starter/zookeeper/autoconfigure/ZkConfigEnvironmentProcessor.java
@@ -20,8 +20,7 @@ package org.dromara.dynamictp.starter.zookeeper.autoconfigure;
 import lombok.extern.slf4j.Slf4j;
 import org.dromara.dynamictp.common.ApplicationContextHolder;
 import org.dromara.dynamictp.common.properties.DtpProperties;
-import org.dromara.dynamictp.core.spring.BinderHelper;
-import org.dromara.dynamictp.core.spring.PropertiesBinder;
+import org.dromara.dynamictp.core.support.BinderHelper;
 import org.dromara.dynamictp.starter.zookeeper.util.CuratorUtil;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.env.EnvironmentPostProcessor;
@@ -31,7 +30,6 @@ import org.springframework.core.env.ConfigurableEnvironment;
 import org.springframework.core.env.MutablePropertySources;
 
 import java.util.Map;
-import java.util.Objects;
 
 /**
  * ZkConfigEnvironmentProcessor related
@@ -47,11 +45,7 @@ public class ZkConfigEnvironmentProcessor implements EnvironmentPostProcessor, O
     @Override
     public void postProcessEnvironment(ConfigurableEnvironment environment, SpringApplication application) {
         DtpProperties dtpProperties = ApplicationContextHolder.getBean(DtpProperties.class);
-        final PropertiesBinder binder = BinderHelper.getBinder();
-        if (Objects.isNull(binder)) {
-            return;
-        }
-        binder.bindDtpProperties(environment, dtpProperties);
+        BinderHelper.bindDtpProperties(environment, dtpProperties);
         Map<Object, Object> properties = CuratorUtil.genPropertiesMap(dtpProperties);
         if (!checkPropertyExist(environment)) {
             createZkPropertySource(environment, properties);
